outsourced AppendBehavior
